Entferne Boxing/Unboxing
IntelliJ sagt:
> Explicit wrapping of primitive values in objects/ manual boxing
> is unnecessary under Java 5 and newer, and can be safely removed.

+++ src/de/willuhn/jameica/hbci/gui/action/Synchronize.java
@@ -199,7 +199,7 @@ public class Synchronize implements Action
         return;
       }
       
-      int status = ((Integer) data).intValue();
+      int status = (Integer) data;
       if (status == ProgressMonitor.STATUS_DONE)
       {
         sync();
