+++ src/de/willuhn/jameica/hbci/gui/parts/UmsatzTypList.java
@@ -70,7 +70,7 @@ public class UmsatzTypList extends TablePart implements Part
       {
         if (o == null)
           return i18n.tr("egal");
-        return UmsatzTypUtil.getNameForType(((Integer) o).intValue());
+        return UmsatzTypUtil.getNameForType((Integer) o);
       }
     });
     addColumn(i18n.tr("Konto"),"dummy");
