+++ src/de/willuhn/jameica/hbci/server/AbstractSepaSammelTransferBuchungImpl.java
@@ -172,7 +172,7 @@ public abstract class AbstractSepaSammelTransferBuchungImpl<T extends SepaSammel
     Double d = (Double) getAttribute("betrag");
     if (d == null)
       return 0;
-    return d.doubleValue();
+    return d;
   }
 
   /**
